Added player actions, a better way to communicate key presses and mouse button presses. All mouse buttons are supported now.

+++ src/main/java/walledin/engine/math/Vector2i.java
@@ -56,7 +56,7 @@ public class Vector2i {
     }
 
     public Vector2f asVector2f() {
-        return new Vector2f((float) x, (float) y);
+        return new Vector2f(x, y);
     }
 
 }
